pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gtd-mcp-server"
version = "0.1.0"
description = "A local MCP server for GTD workflows — Todoist and Gmail integration for Claude"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "fastmcp>=2.0.0",
    "todoist-api-python>=2.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth>=2.25.0",
]

[project.scripts]
gtd-mcp = "gtd_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "black>=24.0",
    "ruff>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gtd_mcp"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
